# Support Files/JSON files/api-docs.json
{"openapi":"3.1.0","info":{"title":"MyAccountingBooks API","description":"Double-entry accounting engine REST API.\n\n## Authentication\n1. Call **POST /auth/login** with your email and password.\n2. Copy the `token` from the response.\n3. Click **Authorize** (top right), paste the token, click **Authorize**.\n4. All subsequent requests will include the Bearer token automatically.\n\n## Rational Arithmetic\nMonetary values use `valueNum / valueDenom` — never floating point.\nExample: MXN $500.00 = `valueNum: 50000`, `valueDenom: 100`.\nUse `GET /commodities/{id}` to retrieve the correct `fraction`\n(which equals `valueDenom`) for any currency.\n\n## RLS Tenant Isolation\nAll ledger, account, and transaction data is scoped to the\nauthenticated owner via PostgreSQL Row-Level Security.\nYou cannot access another owner's data even with a valid token.\n","contact":{"name":"León Felipe Guevara Chávez","email":"leon@myaccountingbooks.app"},"license":{"name":"Proprietary","url":"https://myaccountingbooks.app"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Local development server"}],"security":[{"bearerAuth":[]}],"tags":[{"name":"Auth","description":"Authentication — obtain JWT token"},{"name":"Transactions","description":"Double-entry posting engine — post, reverse, and void transactions"},{"name":"Ledgers","description":"Ledger management — create and list ledgers"},{"name":"Accounts","description":"Chart of Accounts — retrieve account tree for a ledger"},{"name":"Health","description":"Service liveness probe"},{"name":"Commodities","description":"Currency and commodity catalog — ISO-4217 and custom entries"}],"paths":{"/transactions":{"post":{"tags":["Transactions"],"summary":"Post transaction","description":"Posts a balanced double-entry transaction.\n\n**Rational arithmetic:** all monetary values use `valueNum / valueDenom`.\nExample: MXN $500.00 → `valueNum: 50000, valueDenom: 100`.\nAll splits in one request **must share the same `valueDenom`**.\n\n**Balance rule:** `SUM(valueNum WHERE side=0)` must equal\n`SUM(valueNum WHERE side=1)`. Enforced by the database.\n\n**Side values:** `0` = DEBIT, `1` = CREDIT.\n","operationId":"postTransaction","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostTransactionRequest"}}},"required":true},"responses":{"201":{"description":"Transaction posted","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TransactionResponse"}}}},"400":{"description":"Unbalanced splits, wrong accounts, or DB validation failure"},"401":{"description":"Missing or invalid token"}},"security":[{"bearerAuth":[]}]}},"/transactions/{id}/void":{"post":{"tags":["Transactions"],"summary":"Void transaction","description":"Marks a transaction as voided in-place. Does **not** create a new transaction.\nSets `isVoided = true`, `voidedAt = now()`, and appends `[VOID: reason]`\nto the transaction memo if a reason is provided.\n\n**Guards (enforced by database):**\n- Cannot void an already-voided transaction.\n\nSend `{}` if no reason is needed.\n","operationId":"voidTransaction","parameters":[{"name":"id","in":"path","description":"UUID of the transaction to void","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VoidTransactionRequest"}}}},"responses":{"200":{"description":"Transaction voided — isVoided will be true","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TransactionResponse"}}}},"400":{"description":"Transaction is already voided"},"401":{"description":"Missing or invalid token"},"404":{"description":"Transaction not found or not owned by caller"}},"security":[{"bearerAuth":[]}]}},"/transactions/{id}/reverse":{"post":{"tags":["Transactions"],"summary":"Reverse transaction","description":"Creates a mirror reversal transaction with all split sides flipped (DEBIT↔CREDIT).\nMarks the original transaction with `reversedByTxId`.\n\n**Guards (enforced by database):**\n- Cannot reverse a voided transaction.\n- Cannot reverse a deleted transaction.\n- Cannot reverse a transaction that has already been reversed.\n\nAll body fields are optional — send `{}` for a default reversal.\n","operationId":"reverseTransaction","parameters":[{"name":"id","in":"path","description":"UUID of the transaction to reverse","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReverseTransactionRequest"}}}},"responses":{"201":{"description":"Reversal transaction created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TransactionResponse"}}}},"400":{"description":"Transaction already reversed, voided, or deleted"},"401":{"description":"Missing or invalid token"},"404":{"description":"Transaction not found or not owned by caller"}},"security":[{"bearerAuth":[]}]}},"/ledgers":{"get":{"tags":["Ledgers"],"summary":"List ledgers","description":"Returns all active ledgers owned by the authenticated user.","operationId":"getLedgers","responses":{"200":{"description":"List of ledgers","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LedgerResponse"}}}}},"401":{"description":"Missing or invalid token"}},"security":[{"bearerAuth":[]}]},"post":{"tags":["Ledgers"],"summary":"Create ledger","description":"Creates a new ledger for the authenticated owner.\nOptionally instantiates a Chart of Accounts from a template\nby providing `coaTemplateCode` and `coaTemplateVersion`.\n","operationId":"createLedger","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateLedgerRequest"}}},"required":true},"responses":{"201":{"description":"Ledger created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LedgerResponse"}}}},"400":{"description":"Invalid request body"},"401":{"description":"Missing or invalid token"}},"security":[{"bearerAuth":[]}]}},"/auth/login":{"post":{"tags":["Auth"],"summary":"Login","description":"Authenticates with email and password. Returns a JWT token valid for 24 hours.","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Login successful — use the token in all subsequent requests","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"401":{"description":"Invalid credentials"}},"security":[]}},"/ledgers/{ledgerId}/accounts":{"get":{"tags":["Accounts"],"summary":"Get Chart of Accounts","description":"Returns all active accounts for the specified ledger as a **flat list**.\nUse the `parentId` field on each account to reconstruct the tree hierarchy\non the client side.\n\nAccounts are ordered by `code ASC`. Placeholder accounts (grouping nodes\nthat cannot receive transactions) are included and marked with\n`isPlaceholder: true`.\n","operationId":"getAccounts","parameters":[{"name":"ledgerId","in":"path","description":"UUID of the ledger whose accounts to retrieve","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Flat list of accounts","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AccountResponse"}}}}},"401":{"description":"Missing or invalid token"},"404":{"description":"Ledger not found or not owned by caller"}},"security":[{"bearerAuth":[]}]}},"/health":{"get":{"tags":["Health"],"summary":"Health check","description":"Returns OK if the service is running. No authentication required.","operationId":"health","responses":{"200":{"description":"Service is healthy","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}}},"security":[]}},"/commodities":{"get":{"tags":["Commodities"],"summary":"List commodities","description":"Returns all active commodities.\nUse the optional `namespace` parameter to filter by category.\n\n**Primary use case:** fetch `GET /commodities?namespace=CURRENCY`\nto populate a currency picker. The `id` from each result is used\nas `currencyCommodityId` when creating a ledger or posting a transaction.\n\nThe `fraction` field equals the correct `valueDenom` to use in splits\nfor that currency. Example: MXN `fraction=100` → use `valueDenom: 100`.\n","operationId":"getCommodities","parameters":[{"name":"namespace","in":"query","description":"Optional namespace filter. Valid values: CURRENCY, STOCK, FUND, INDEX, OTHER","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"List of active commodities","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CommodityResponse"}}}}},"400":{"description":"Invalid namespace value"},"401":{"description":"Missing or invalid token"}},"security":[{"bearerAuth":[]}]}},"/commodities/{id}":{"get":{"tags":["Commodities"],"summary":"Get commodity by ID","description":"Returns a single active commodity by its UUID primary key.","operationId":"getCommodityById","parameters":[{"name":"id","in":"path","description":"UUID of the commodity","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Commodity found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommodityResponse"}}}},"401":{"description":"Missing or invalid token"},"404":{"description":"Commodity not found or inactive"}},"security":[{"bearerAuth":[]}]}}},"components":{"schemas":{"PostTransactionRequest":{"type":"object","properties":{"ledgerId":{"type":"string","format":"uuid"},"currencyCommodityId":{"type":"string","format":"uuid"},"postDate":{"type":"string","format":"date-time"},"enterDate":{"type":"string","format":"date-time"},"memo":{"type":"string"},"num":{"type":"string"},"status":{"type":"integer","format":"int32"},"payeeId":{"type":"string","format":"uuid"},"splits":{"type":"array","items":{"$ref":"#/components/schemas/SplitRequest"},"maxItems":2147483647,"minItems":2}},"required":["currencyCommodityId","ledgerId","splits"]},"SplitRequest":{"type":"object","properties":{"accountId":{"type":"string","format":"uuid"},"side":{"type":"integer","format":"int32"},"valueNum":{"type":"integer","format":"int64"},"valueDenom":{"type":"integer","format":"int32"},"quantityNum":{"type":"integer","format":"int64"},"quantityDenom":{"type":"integer","format":"int32"},"memo":{"type":"string"},"action":{"type":"string"}},"required":["accountId"]},"SplitResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"accountId":{"type":"string","format":"uuid"},"side":{"type":"integer","format":"int32"},"valueNum":{"type":"integer","format":"int64"},"valueDenom":{"type":"integer","format":"int32"},"memo":{"type":"string"}}},"TransactionResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"ledgerId":{"type":"string","format":"uuid"},"currencyCommodityId":{"type":"string","format":"uuid"},"postDate":{"type":"string","format":"date-time"},"enterDate":{"type":"string","format":"date-time"},"memo":{"type":"string"},"num":{"type":"string"},"isVoided":{"type":"boolean"},"splits":{"type":"array","items":{"$ref":"#/components/schemas/SplitResponse"}}}},"VoidTransactionRequest":{"type":"object","properties":{"reason":{"type":"string"}}},"ReverseTransactionRequest":{"type":"object","properties":{"postDate":{"type":"string","format":"date-time"},"enterDate":{"type":"string","format":"date-time"},"memo":{"type":"string"}}},"CreateLedgerRequest":{"type":"object","properties":{"name":{"type":"string","minLength":1},"currencyMnemonic":{"type":"string","minLength":1},"decimalPlaces":{"type":"integer","format":"int32","maximum":8,"minimum":0},"coaTemplateCode":{"type":"string"},"coaTemplateVersion":{"type":"string"}}},"LedgerResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"currencyCode":{"type":"string"},"decimalPlaces":{"type":"integer","format":"int32"}}},"LoginRequest":{"type":"object","properties":{"email":{"type":"string","minLength":1},"password":{"type":"string","minLength":1}}},"TokenResponse":{"type":"object","properties":{"token":{"type":"string"},"ownerID":{"type":"string","format":"uuid"}}},"AccountResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"code":{"type":"string"},"parentId":{"type":"string","format":"uuid"},"isPlaceholder":{"type":"boolean"},"isHidden":{"type":"boolean"},"kind":{"type":"integer","format":"int32"},"accountTypeCode":{"type":"string"}}},"CommodityResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"mnemonic":{"type":"string"},"namespace":{"type":"string"},"fullName":{"type":"string"},"fraction":{"type":"integer","format":"int64"},"isActive":{"type":"boolean"}}}},"securitySchemes":{"bearerAuth":{"type":"http","description":"Paste the JWT token returned by POST /auth/login. Format: Bearer <token> (the 'Bearer ' prefix is added automatically).","scheme":"bearer","bearerFormat":"JWT"}}}}